// webrtc/server-sdk/api.ts
export interface AddComponentRequest {
  /**
   *
   * @type {ComponentOptions}
   * @memberof AddComponentRequest
   */
  options?: ComponentOptions;
  /**
   * Component type
   * @type {string}
   * @memberof AddComponentRequest
   */
  type: string;
}
/**
 *
 * @export
 * @interface AddPeerRequest
 */
export interface AddPeerRequest {
  /**
   *
   * @type {PeerOptions}
   * @memberof AddPeerRequest
   */
  options: PeerOptions;
  /**
   * Peer type
   * @type {string}
   * @memberof AddPeerRequest
   */
  type: string;
}
// ...
export type ComponentOptions =
  | ComponentOptionsFile
  | ComponentOptionsHLS
  | ComponentOptionsRTSP
  | ComponentOptionsRecording
  | ComponentOptionsSIP;
// ...
export type PeerOptions = PeerOptionsWebRTC;
// ...
export interface RoomConfig {
  /**
   * Maximum amount of peers allowed into the room
   * @type {number}
   * @memberof RoomConfig
   */
  maxPeers?: number | null;
  /**
   * Duration (in seconds) after which the peer will be removed if it is disconnected. If not provided, this feature is disabled.
   * @type {number}
   * @memberof RoomConfig
   */
  peerDisconnectedTimeout?: number | null;
  /**
   * Duration (in seconds) after which the room will be removed if no peers are connected. If not provided, this feature is disabled.
   * @type {number}
   * @memberof RoomConfig
   */
  peerlessPurgeTimeout?: number | null;
  /**
   * Custom id used for identifying room within Fishjam. Must be unique across all rooms. If not provided, random UUID is generated.
   * @type {string}
   * @memberof RoomConfig
   */
  roomId?: string | null;
  /**
   * Enforces video codec for each peer in the room
   * @type {string}
   * @memberof RoomConfig
   */
  videoCodec?: RoomConfigVideoCodecEnum | null;
  /**
   * URL where Fishjam notifications will be sent
   * @type {string}
   * @memberof RoomConfig
   */
  webhookUrl?: string | null;
}

export const RoomConfigVideoCodecEnum = {
  H264: "h264",
  Vp8: "vp8",
} as const;

export type RoomConfigVideoCodecEnum =
  (typeof RoomConfigVideoCodecEnum)[keyof typeof RoomConfigVideoCodecEnum];
// ...
export interface SubscriptionConfig {
  /**
   * List of peers and components ids whose tracks the HLS endpoint will subscribe to
   * @type {Array<string>}
   * @memberof SubscriptionConfig
   */
  origins?: Array<string>;
}
// ...
export class RoomApi extends BaseAPI {
  /**
   *
   * @summary Creates the component and adds it to the room
   * @param {string} roomId Room ID
   * @param {AddComponentRequest} [addComponentRequest] Component config
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof RoomApi
   */
  public addComponent(
    roomId: string,
    addComponentRequest?: AddComponentRequest,
    options?: RawAxiosRequestConfig,
  ) {
    return RoomApiFp(this.configuration)
      .addComponent(roomId, addComponentRequest, options)
      .then((request) => request(this.axios, this.basePath));
  }

  /**
   *
   * @summary Create peer
   * @param {string} roomId Room id
   * @param {AddPeerRequest} [addPeerRequest] Peer specification
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof RoomApi
   */
  public addPeer(
    roomId: string,
    addPeerRequest?: AddPeerRequest,
    options?: RawAxiosRequestConfig,
  ) {
    return RoomApiFp(this.configuration)
      .addPeer(roomId, addPeerRequest, options)
      .then((request) => request(this.axios, this.basePath));
  }

  /**
   *
   * @summary Creates a room
   * @param {RoomConfig} [roomConfig] Room configuration
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof RoomApi
   */
  public createRoom(roomConfig?: RoomConfig, options?: RawAxiosRequestConfig) {
    return RoomApiFp(this.configuration)
      .createRoom(roomConfig, options)
      .then((request) => request(this.axios, this.basePath));
  }

  /**
   *
   * @summary Delete the component from the room
   * @param {string} roomId Room ID
   * @param {string} id Component ID
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof RoomApi
   */
  public deleteComponent(
    roomId: string,
    id: string,
    options?: RawAxiosRequestConfig,
  ) {
    return RoomApiFp(this.configuration)
      .deleteComponent(roomId, id, options)
      .then((request) => request(this.axios, this.basePath));
  }

  /**
   *
   * @summary Delete peer
   * @param {string} roomId Room ID
   * @param {string} id Peer id
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof RoomApi
   */
  public deletePeer(
    roomId: string,
    id: string,
    options?: RawAxiosRequestConfig,
  ) {
    return RoomApiFp(this.configuration)
      .deletePeer(roomId, id, options)
      .then((request) => request(this.axios, this.basePath));
  }

  /**
   *
   * @summary Delete the room
   * @param {string} roomId Room id
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof RoomApi
   */
  public deleteRoom(roomId: string, options?: RawAxiosRequestConfig) {
    return RoomApiFp(this.configuration)
      .deleteRoom(roomId, options)
      .then((request) => request(this.axios, this.basePath));
  }

  /**
   *
   * @summary Show information about all rooms
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof RoomApi
   */
  public getAllRooms(options?: RawAxiosRequestConfig) {
    return RoomApiFp(this.configuration)
      .getAllRooms(options)
      .then((request) => request(this.axios, this.basePath));
  }

  /**
   *
   * @summary Shows information about the room
   * @param {string} roomId Room ID
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof RoomApi
   */
  public getRoom(roomId: string, options?: RawAxiosRequestConfig) {
    return RoomApiFp(this.configuration)
      .getRoom(roomId, options)
      .then((request) => request(this.axios, this.basePath));
  }

  /**
   *
   * @summary Subscribe component to the tracks of peers or components
   * @param {string} roomId Room ID
   * @param {string} componentId Component ID
   * @param {SubscriptionConfig} [subscriptionConfig] Subscribe configuration
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof RoomApi
   */
  public subscribeTo(
    roomId: string,
    componentId: string,
    subscriptionConfig?: SubscriptionConfig,
    options?: RawAxiosRequestConfig,
  ) {
    return RoomApiFp(this.configuration)
      .subscribeTo(roomId, componentId, subscriptionConfig, options)
      .then((request) => request(this.axios, this.basePath));
  }
}